+++ src/ts/Components/Atoms/Layout/Layout.ts
@@ -1,3 +1,6 @@
+/**
+ * @module Atoms/Layout
+ */
 import { iAnaConfiguration } from '../../../Ana/Ana.interface'
 import { LayoutMargin, LayoutType, RenderDictionary } from '../../../types'
 import classNames from 'classnames'
@@ -12,7 +15,7 @@ import { LayoutClass } from '../../Particles/Particles'
 //              |___/
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
 /**
- * This is an Ana Component, it is not used to render the HTMLElement itself. Instead, this class is used to create a render function based on the global configuration.
+ * 
  */
 export function rLayout(
   a: RenderDictionary,