+++ src/ts/Components/Atoms/Paragraph/Paragraph.ts
@@ -1,31 +1,47 @@
-//    ___                                  _    
-//   | _ \__ _ _ _ __ _ __ _ _ _ __ _ _ __| |_  
-//   |  _/ _` | '_/ _` / _` | '_/ _` | '_ \ ' \ 
-//   |_| \__,_|_| \__,_\__, |_| \__,_| .__/_||_|
-//                     |___/         |_|        
+/**
+ * @module Atoms/Paragraph
+ */
+import { iAnaConfiguration } from '../../../Ana/Ana.interface'
+import { RenderDictionary, ContrastValues, ColorValues } from '../../../types'
+import classNames from 'classnames'
+import { ColorContrastClass } from '../../Particles/Particles'
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+//  ____                                       _
+// |  _ \ __ _ _ __ __ _  __ _ _ __ __ _ _ __ | |__
+// | |_) / _` | '__/ _` |/ _` | '__/ _` | '_ \| '_ \
+// |  __/ (_| | | | (_| | (_| | | | (_| | |_) | | | |
+// |_|   \__,_|_|  \__,_|\__, |_|  \__,_| .__/|_| |_|
+//                       |___/          |_|
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
 /**
- * @module Component/Atom/Paragraph
+ * A Paragraph is one of two base typographical elements, slightly larger than a Label. This element is used for text that has multiple lines and more text after it, maybe a heading or another paragraph. If the content inside the element is a few words long, use a Label instead.
  */
- import { AnaComponent } from '../../Components'
+export function rParagraph(
+  a: RenderDictionary,
+  config: iAnaConfiguration
+): Function {
+  return (...children: [Node | string | Function]): Function => {
+    return (param: iParagraph = {}): HTMLElement => {
+      // Default values
+      param = {
+        ...{ contrast: 'highest', color: 'grey' },
+        ...param,
+      }
+      let classes = {
+        Paragraph: classNames(
+          'a-Paragraph',
+          ColorContrastClass(param.color, param.contrast, 'txt')
+        ).split(' '),
+      }
 
- /**
-  * A Paragraph is one of two base typographical elements, slightly larger than a Label. This element is used for text that has multiple lines and more text after it, maybe a heading or another paragraph. If the content inside the element is a few words long, use a Label instead.
-  */
- export class Paragraph extends AnaComponent {
-   render: Function = (param: iParagraph = {}): HTMLElement => {
-     const a = this.a
-     //param = extendParameterDefaults<tParagraph>(param, { text: 'Paragraph' })
- 
-     return a
-       .div(a.span(param.text).has({ class: this.classRoot }))
-       .has({ class: `${this.classRoot}-container` })
-   }
- }
- 
- interface iParagraph {
-   text?: string
-   [key: string]: tParagraph
- }
- 
- type tParagraph = undefined | string
+      return a.div(...classes.Paragraph)(...children)
+    }
+  }
+}
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+export interface iParagraph {
+  contrast?: ContrastValues
+  color?: ColorValues
+}