+++ src/ts/Components/Atoms/Subheading/Subheading.ts
@@ -1,31 +1,47 @@
-//    ___      _    _                _ _           
-//   / __|_  _| |__| |_  ___ __ _ __| (_)_ _  __ _ 
-//   \__ \ || | '_ \ ' \/ -_) _` / _` | | ' \/ _` |
-//   |___/\_,_|_.__/_||_\___\__,_\__,_|_|_||_\__, |
-//                                           |___/ 
+ /**
+ * @module Atoms/Subheading
+ */
+import { iAnaConfiguration } from '../../../Ana/Ana.interface'
+import { RenderDictionary, ContrastValues, ColorValues } from '../../../types'
+import classNames from 'classnames'
+import { ColorContrastClass } from '../../Particles/Particles'
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+//   ____        _     _                    _ _             
+//  / ___| _   _| |__ | |__   ___  __ _  __| (_)_ __   __ _ 
+//  \___ \| | | | '_ \| '_ \ / _ \/ _` |/ _` | | '_ \ / _` |
+//   ___) | |_| | |_) | | | |  __/ (_| | (_| | | | | | (_| |
+//  |____/ \__,_|_.__/|_| |_|\___|\__,_|\__,_|_|_| |_|\__, |
+//                                                    |___/ 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
 /**
- * @module Component/Atom/Subheading
+ * A Subheading is the smallest of the three heading types, along with a Heading and Title. It is conventionally used for titles of subsections, multiple of these can be found inside a section.
  */
- import { AnaComponent } from '../../Components'
+export function rSubheading(
+  a: RenderDictionary,
+  config: iAnaConfiguration
+): Function {
+  return (...children: [Node | string | Function]): Function => {
+    return (param: iSubheading = {}): HTMLElement => {
+      // Default values
+      param = {
+        ...{ contrast: 'highest', color: 'grey' },
+        ...param,
+      }
+      let classes = {
+        Subheading: classNames(
+          'a-Subheading',
+          ColorContrastClass(param.color, param.contrast, 'txt')
+        ).split(' '),
+      }
 
- /**
-  * A Subheading is the smallest of the three heading types, the other two being the Heading and Title. It is conventionally used for titles of subsections, multiple of these can be found inside a section.
-  */
- export class Subheading extends AnaComponent {
-   render: Function = (param: iSubheading = {}): HTMLElement => {
-     const a = this.a
-     //param = extendParameterDefaults<tSubheading>(param, { text: 'Subheading' })
- 
-     return a
-       .div(a.span(param.text).has({ class: this.classRoot }))
-       .has({ class: `${this.classRoot}-container` })
-   }
- }
- 
- interface iSubheading {
-   text?: string
-   [key: string]: tSubheading
- }
- 
- type tSubheading = undefined | string
+      return a.div(...classes.Subheading)(...children)
+    }
+  }
+}
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+export interface iSubheading {
+  contrast?: ContrastValues
+  color?: ColorValues
+}